nv_fill_rectangle: add test
This adds a basic shader test for the GL_NV_fill_rectangle extension.
The test simply draws a single prim with the polygon mode set to
GL_FILL_RECTANGLE_NV and checks that the prim was rasterized as a full
rectangle, along with checking that the rectangle only occupies the
space we gave it in it's vertex data.

Changes since v1:
- Use GL_TRIANGLES in basic test
- Use green for prims, 0.2 for clear
- Fix points used for relative probe rect rgb

Signed-off-by: Lyude <lyude@redhat.com>